+++ demo/src/app/gridster/IGridsterOptions.ts
@@ -5,4 +5,11 @@ export interface IGridsterOptions {
     heightToFontSizeRatio?: number;
     dragAndDrop?: boolean;
     itemSelector?: string;
+    resizable?: boolean;
+    minWidth?: number;
+    minHeight?: number;
+    maxWidth?: number;
+    maxHeight?: number;
+    defaultItemWidth?: number;
+    defaultItemHeight?: number;
 }